% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{get_metric}
\alias{get_metric}
\title{Helper function to compute metrics from a list of models summmaries.}
\usage{
get_metric(model_sum, metric)
}
\arguments{
\item{model_sum}{list of model summary objects}

\item{metric}{metric from summary to return}
}
\value{
A numeric vector
}
\description{
Helper function to compute metrics from a list of models summmaries.
}
\examples{
# internal use only
}
