% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{metric_calc}
\alias{metric_calc}
\title{Helper function to compute metrics from a list of models summmaries.}
\usage{
metric_calc(model_sum, metric)
}
\arguments{
\item{model_sum}{list of model summary objects}

\item{metric}{desired metric to calculate}
}
\value{
A numeric vector
}
\description{
Helper function to compute metrics from a list of models summmaries.
}
\examples{
# internal use only
}
