% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{smelt}
\alias{smelt}
\title{A function to combine models.}
\usage{
smelt(...)
}
\arguments{
\item{...}{list of models to summarize.}
}
\value{
output
}
\description{
A function to combine models.
}
\examples{
lm_1 = lm(mpg ~ cyl + disp + hp, data = mtcars)
lm_2 = lm(mpg ~ cyl + wt + drat, data = mtcars)
smelt(lm_1, lm_2)
}
