% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{stack_coeff}
\alias{stack_coeff}
\title{With several models, we want to compare their coefficients, confidence
intervals, and standard errors.}
\usage{
stack_coeff(...)
}
\arguments{
\item{...}{A list of models to summarize and combine.}
}
\value{
A \code{\link[=data.frame]{data.frame()}}
}
\description{
With several models, we want to compare their coefficients, confidence
intervals, and standard errors.
}
